explicitly convert `MathLib::big{u}int` to/from other types
fixes some compiler warnings with `USE_BOOST_INT128` as those implicit conversions are not allowed

+++ lib/programmemory.cpp
@@ -575,7 +575,7 @@ static bool isNumericValue(const ValueFlow::Value& value) {
 
 static double asFloat(const ValueFlow::Value& value)
 {
-    return value.isFloatValue() ? value.floatValue : value.intvalue;
+    return value.isFloatValue() ? value.floatValue : static_cast<double>(value.intvalue);
 }
 
 static std::string removeAssign(const std::string& assign) {
@@ -1467,7 +1467,7 @@ namespace {
                     return unknown();
                 const MathLib::bigint index = rhs.intvalue;
                 if (index >= 0 && index < strValue.size())
-                    return ValueFlow::Value{strValue[index]};
+                    return ValueFlow::Value{strValue[static_cast<std::size_t>(index)]};
                 if (index == strValue.size())
                     return ValueFlow::Value{};
             } else if (Token::Match(expr, "%cop%") && expr->astOperand1() && expr->astOperand2()) {
